# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для генерации персонализированных поздравлений: AI-изображения + песни. Аутентификация через httpOnly cookie (access_token).",
        "title": "FunBaza API",
        "contact": {
            "name": "FunBaza Team"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/billing/balance": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает текущий баланс кредитов авторизованного пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Баланс кредитов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.BalanceResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/billing/estimate": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Считает стоимость генерации до её запуска. Используется для отображения цены пользователю.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Предварительный расчёт стоимости",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Количество изображений (0-3)",
                        "name": "images",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Количество песен (0-3)",
                        "name": "songs",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.EstimateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/billing/tariff": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает активный тариф с ценами за изображение и песню.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Текущий тариф",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Tariff"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/billing/transactions": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает постраничный список кредитных транзакций пользователя (начисления, списания, возвраты).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "История транзакций",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит записей (макс. 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TransactionsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.BalanceResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "handlers.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "message": {
                    "type": "string",
                    "example": "Refresh token missing"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handlers.ErrorDetail"
                }
            }
        },
        "handlers.EstimateResponse": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "integer",
                    "example": 70
                },
                "images": {
                    "type": "integer",
                    "example": 3
                },
                "price_per_image": {
                    "type": "integer",
                    "example": 15
                },
                "price_per_song": {
                    "type": "integer",
                    "example": 25
                },
                "songs": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "handlers.TransactionsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CreditTransaction"
                    }
                }
            }
        },
...
        "models.CreditTransaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "положительное = начисление, отрицательное = списание",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "reference_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.Tariff": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price_per_image": {
                    "type": "integer"
                },
                "price_per_lyrics": {
                    "type": "integer"
                },
                "price_per_song": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}